Allow using cmd-key combos and escape key in select mode
Github (and other sites) use cmd-k or similar combos to bring up
full-screen command palettes. Transparent binding mode should handle
this, but it doesn't pay attention to modifier keys and therefore a
binding to 'k' will block 'cmd-k' as well.

https://github.com/ccampbell/mousetrap/issues/441 as well as some local
changes to how `stopSitePropagation` works could potentially solve this.

For now let's just assume no one is going to bind viamri commands to cmd
and let it always work.

This also allows escape keys to propagate to the underlying site if
already in select mode, which didn't work before (making it actually
impossible to send an 'escape' key to the site).